requirements.txt
# -------- Core Audio I/O --------
sounddevice==0.4.6
pyaudio==0.2.13

# -------- Speech-to-Text (ASR) --------
openai-whisper==20231117
torch==2.1.2
torchaudio==2.1.2
numpy>=1.22.0,<1.23.0

# -------- LLM Inference --------
llama-cpp-python==0.2.25

# -------- Text-to-Speech (TTS) --------
TTS==0.22.0

# -------- Backend / Control Plane --------
fastapi==0.111.0
uvicorn==0.30.1
websockets==12.0

# -------- Memory / Embeddings --------
sentence-transformers==2.7.0
#chromadb==0.4.24
faiss-cpu==1.8.0

# -------- Vision (optional but safe) --------
opencv-python==4.9.0.80
mediapipe==0.10.14

# -------- Utilities --------
pydantic==2.6.4
python-dotenv==1.0.1
psutil==5.9.8
requests==2.32.3
rich==13.7.1
tqdm==4.66.4
